refactor(api-keys): extract EnrichedApiKey model to dedicated file
Improve code organization by moving the EnrichedApiKey type definition
from the adapter to a dedicated models file, following single responsibility
principle and making the type more discoverable.
