+++ ui/components/users/profile/edit-api-key-name-modal.tsx
@@ -5,7 +5,7 @@ import { ModalFooter } from "@heroui/modal";
 import { useEffect } from "react";
 
 import { updateApiKey } from "@/actions/api-keys/api-keys";
-import { type EnrichedApiKey } from "@/actions/api-keys/api-keys.adapter";
+import { type EnrichedApiKey } from "@/actions/api-keys/models";
 import { Alert, AlertDescription } from "@/components/ui/alert/Alert";
 import { CustomAlertModal } from "@/components/ui/custom/custom-alert-modal";
 import { CustomButton } from "@/components/ui/custom/custom-button";
